requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-text-splitters
langgraph
sentence-transformers
chromadb
pypdf
python-dotenv